ml_dashboard/lang_chain_agentics/langgraph/requirements.txt
# openai
# langchain
# streamlit

streamlit
pyngrok
google-generativeai
